Finish updating to proj-props.  Set warning level to 4, fixed a few of the worst offending errors.

+++ aws_iot_src/shadow/aws_iot_shadow_interface.h
@@ -49,7 +49,7 @@ typedef struct {
 	char *pMyThingName; ///< Every device has a Thing Shadow and this is the placeholder for name
 	char *pMqttClientId; ///< Currently the Shadow uses MQTT to connect and it is important to ensure we have unique client id
 	char *pHost; ///< This will be unique to a customer and can be retrieved from the console
-	int port; ///< By default the port is 8883
+	uint16_t port; ///< By default the port is 8883
 	char *pRootCA; ///< Location with the Filename of the Root CA
 	char *pClientCRT; ///< Location of Device certs signed by AWS IoT service
 	char *pClientKey; ///< Location of Device private key
